requirements.txt
bidict==0.22.1
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
PyQt6==6.5.0
PyQt6-Qt6==6.5.0
PyQt6-sip==13.5.0
pyqtdarktheme==2.1.0
python-engineio==4.4.0
python-socketio==5.8.0
requests==2.28.2
urllib3==1.26.15
websocket-client==1.5.1
cryptography
pytest-playwright==0.3.2
localStoragePy==0.2.3
uvicorn==0.21.1